feat/refactor: add checkout page and refactor the cart, register and home pages
