+++ src/components/require-auth/index.tsx
@@ -11,6 +11,6 @@ export function RequireAuth() {
       <Outlet />
     </>
   ) : (
-    <Navigate to="/" state={{ from: location }} replace />
+    <Navigate to="/login" state={{ from: location }} replace />
   );
 }
