+++ src/pages/checkout/checkout.style.css
@@ -1,0 +1,228 @@
+/* ===========================
+   MAIN LAYOUT
+=========================== */
+.checkout-page {
+    padding: 2rem 0;
+
+}
+
+.checkout-container {
+  max-width: 1200px;
+  margin: 0 auto;
+  display: grid;
+  grid-template-columns: 2fr 1fr;
+  gap: 2rem;
+  align-items: flex-start;
+}
+
+/* ===========================
+   CHECKOUT STEPS
+=========================== */
+.checkout-steps {
+  display: flex;
+  justify-content: space-around;
+  margin-bottom: 2rem;
+  padding: 1rem;
+  background-color: #fff;
+  border-radius: 8px;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+}
+
+.step {
+  display: flex;
+  align-items: center;
+  color: #888;
+  font-weight: 500;
+  padding: 0.5rem 1rem;
+  border-radius: 20px;
+  transition: all 0.3s ease;
+}
+
+.step.active {
+  color: #d9534f;
+  font-weight: 700;
+}
+
+.step .step-icon {
+  margin-right: 0.5rem;
+  font-size: 1.2rem;
+}
+
+/* ===========================
+   CARD STYLES
+=========================== */
+.checkout-card {
+  background-color: #fff;
+  border-radius: 8px;
+  padding: 1.5rem;
+  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.05);
+  margin-bottom: 1.5rem;
+}
+
+.checkout-card h3 {
+  font-size: 1.5rem;
+  margin-bottom: 1.5rem;
+  color: #333;
+  display: flex;
+  align-items: center;
+  border-bottom: 1px solid #eee;
+  padding-bottom: 1rem;
+}
+
+.checkout-card h3 .pi {
+  margin-right: 0.75rem;
+  color: #d9534f;
+}
+
+/* ===========================
+   ORDER ITEMS
+=========================== */
+.order-item-list {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+}
+
+.order-item {
+  display: flex;
+  align-items: center;
+  gap: 1rem;
+}
+
+.order-item-image img {
+  width: 80px;
+  height: 80px;
+  object-fit: cover;
+  border-radius: 4px;
+  border: 1px solid #eee;
+}
+
+.order-item-details {
+  flex: 1;
+}
+
+.order-item-name {
+  font-weight: 600;
+  color: #333;
+  margin: 0 0 0.25rem;
+}
+
+.order-item-meta {
+  font-size: 0.85rem;
+  color: #777;
+}
+
+.order-item-price {
+  font-weight: 600;
+  color: #333;
+  text-align: right;
+}
+
+.order-item-quantity {
+  font-size: 0.9rem;
+  color: #555;
+}
+
+/* ===========================
+   SHIPPING & PAYMENT
+=========================== */
+.address-display, .payment-display {
+  font-size: 1rem;
+  color: #555;
+  line-height: 1.6;
+}
+
+.address-display p, .payment-display p {
+  margin: 0;
+}
+
+.address-display strong {
+  display: block;
+  margin-bottom: 0.25rem;
+  color: #333;
+}
+
+.btn-change {
+  margin-top: 1rem;
+  font-size: 0.9rem;
+}
+
+/* ===========================
+   SUMMARY
+=========================== */
+.summary-sticky {
+  position: sticky;
+  top: 2rem;
+}
+
+.summary-line {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 0.75rem;
+  font-size: 1rem;
+  color: #555;
+}
+
+.summary-line span:last-child {
+  font-weight: 600;
+  color: #333;
+}
+
+.summary-divider {
+  height: 1px;
+  background-color: #eee;
+  margin: 1.5rem 0;
+}
+
+.summary-total {
+  display: flex;
+  justify-content: space-between;
+  font-size: 1.25rem;
+  font-weight: 700;
+  color: #333;
+}
+
+.total-value {
+  color: #d9534f;
+}
+
+.btn-place-order {
+  width: 100%;
+  padding: 1rem;
+  font-size: 1.1rem;
+  margin-top: 1.5rem;
+}
+
+/* ===========================
+   RESPONSIVENESS
+=========================== */
+@media (max-width: 992px) {
+  .checkout-container {
+    grid-template-columns: 1fr;
+  }
+}
+
+@media (max-width: 768px) {
+  .checkout-page {
+    padding: 2rem 1rem;
+  }
+
+  .checkout-card {
+    padding: 1rem;
+  }
+
+  .checkout-card h3 {
+    font-size: 1.25rem;
+  }
+
+  .order-item {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 0.5rem;
+  }
+
+  .order-item-price {
+    text-align: left;
+    margin-top: 0.5rem;
+  }
+}
